Change installation script (#225, #227) (#236)
Change the installation script so that it setups bash in .bashrc and not
in .bash_profile. Note that this leaves users that have previously
installed with sourcing in two places - this is acceptable as it does
not bring harm, except wasting CPU on parsing the script twice.

Also change the way files are include (both in bash and in zsh), adding
guards if the file exists. This avoids error messages if user at some
point deletes $YVM_DIR.